Separate helper functions.
